refactor: extract services from GroupGenService following SRP
- Create KeywordExtractionService for keyword extraction logic with coroutines
- Create GroupingService for grouping logic and validation
- Create GroupContentGenerationService for content generation
- Create DateTimeService for time-related utilities
- Refactor GroupGenService to orchestrate extracted services
- Reduce GroupGenService from 322 to 144 lines
- Improve separation of concerns and testability
